requirements.txt
Django == 3.1.2
pytest
pytest-django
pytest-cov
factory-boy
responses
requests
celery==5.0.1
redis==3.5.3
python-dotenv
vcgencmd
pyserial
django-celery
gpiozero
RPi.GPIO
smbus
telebot
pytelegrambotapi
viberbot==1.0.11
django-crispy-forms
speedtest-cli
djangorestframework==3.13.1

matplotlib
